requirements.txt
python-binance
binance-connector
python-dotenv
apscheduler
requests
tzlocal
urllib3
websocket-client
